Update trashcan collision to match new appearance Add loading overlay, disabled for now.

+++ common/src/main/java/com/unlikepaladin/pfm/runtime/PFMResourceProgress.java
@@ -1,7 +1,7 @@
 package com.unlikepaladin.pfm.runtime;
 
 public interface PFMResourceProgress {
-    int getProgress();
+    float getProgress();
 
     String getProgressString();
 }
